fixes undo
