+++ app/src/main/java/com/game/prayansh/ultimatetictactoe/models/Game.java
@@ -92,9 +92,11 @@ public class Game {
     public Move playMove(int block, int cell) throws InvalidMoveException, GameOverException, InvalidBlockException, BoardSolvedException {
         CellVal player = getPlayer();
         int contextIndex = getContextBoardIndex();
+        boolean freeHit = false;
 
         if (contextIndex == -1) {
             setContextBoardIndex(block);
+            freeHit = true;
         } else if (contextIndex != block) {
             throw new InvalidBlockException("Found " + contextIndex + " but need to play at " + block, contextIndex, block);
         }
@@ -111,7 +113,7 @@ public class Game {
         contextIndex = getContextBoardIndex();
 
         boolean valid = getContextGameBoard().setCellAt(cell, player);
-        Move m = new Move(getContextBoardIndex(), cell, moves.getFlag());
+        Move m = new Move(getContextBoardIndex(), cell, freeHit);
         if (!valid || !moves.push(m)) {
             throw new InvalidMoveException("Invalid Move for Player " + player.name() + ":" + cell);
         }
